+++ src/Command/ProcessBounceMailsCommand.php
@@ -3,7 +3,6 @@
 namespace Foodsharing\Command;
 
 use Foodsharing\Lib\Mail\BounceProcessing;
-use Foodsharing\Modules\Core\InfluxMetrics;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -13,12 +12,10 @@ class ProcessBounceMailsCommand extends Command
     protected static $defaultName = 'foodsharing:process-bounce-emails';
 
     private $bounceProcessing;
-    private $influxMetrics;
 
-    public function __construct(BounceProcessing $bounceProcessing, InfluxMetrics $influxMetrics)
+    public function __construct(BounceProcessing $bounceProcessing)
     {
         $this->bounceProcessing = $bounceProcessing;
-        $this->influxMetrics = $influxMetrics;
         parent::__construct();
     }
 
@@ -30,7 +27,6 @@ class ProcessBounceMailsCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $this->bounceProcessing->process();
-        $this->influxMetrics->addPoint('mail_bounce', [], ['cnt' => $this->bounceProcessing->getNumberOfProcessedBounces()]);
 
         return 0;
     }